Split db/models.py into one model per file
