+++ convergence_games/db/models/__init__.py
@@ -1,0 +1,69 @@
+# `LoginProvider` is not a model, but the pre-split `models.py` imported it from `db.enums` at
+# module scope, so downstream code (e.g. accounts/controllers/_oauth.py) does
+# `from convergence_games.db.models import LoginProvider`. Re-export it here to keep that working.
+from convergence_games.db.enums import LoginProvider
+
+from ._allocation import Allocation
+from ._base import Base, UserAuditColumns, foreign_key_constraint_with_event
+from ._content_warning import ContentWarning
+from ._event import Event
+from ._game import Game
+from ._game_content_warning_link import GameContentWarningLink
+from ._game_genre_link import GameGenreLink
+from ._game_image_link import GameImageLink
+from ._game_requirement import GameRequirement
+from ._game_requirement_time_slot_link import GameRequirementTimeSlotLink
+from ._genre import Genre
+from ._image import Image
+from ._party import Party
+from ._party_user_link import PartyUserLink
+from ._room import Room
+from ._session import Session
+from ._system import System
+from ._system_alias import SystemAlias
+from ._table import Table
+from ._time_slot import TimeSlot
+from ._user import User
+from ._user_checkin_status import UserCheckinStatus
+from ._user_email_verification_code import UserEmailVerificationCode
+from ._user_event_compensation_transaction import UserEventCompensationTransaction
+from ._user_event_d20_transaction import UserEventD20Transaction
+from ._user_event_role import UserEventRole
+from ._user_game_played import UserGamePlayed
+from ._user_game_preference import UserGamePreference
+from ._user_login import UserLogin
+
+__all__ = [
+    "Allocation",
+    "Base",
+    "ContentWarning",
+    "Event",
+    "Game",
+    "GameContentWarningLink",
+    "GameGenreLink",
+    "GameImageLink",
+    "GameRequirement",
+    "GameRequirementTimeSlotLink",
+    "Genre",
+    "Image",
+    "LoginProvider",
+    "Party",
+    "PartyUserLink",
+    "Room",
+    "Session",
+    "System",
+    "SystemAlias",
+    "Table",
+    "TimeSlot",
+    "User",
+    "UserAuditColumns",
+    "UserCheckinStatus",
+    "UserEmailVerificationCode",
+    "UserEventCompensationTransaction",
+    "UserEventD20Transaction",
+    "UserEventRole",
+    "UserGamePlayed",
+    "UserGamePreference",
+    "UserLogin",
+    "foreign_key_constraint_with_event",
+]
